examples/serialvector/python/requirements.txt
HersheyFonts
pygame
pyserial
